Feature: PHP 8.3 support

+++ pipe.sh
@@ -33,40 +33,44 @@ SQL
 }
 
 composer_setup () {
-  if [ ! -f "composer.lock" ]; then
-      echo "composer.lock does not exist."
-      composer create-project --repository-url="$REPOSITORY_URL" "$MAGENTO_VERSION" /magento2 --no-install
-      cd /magento2
-
-    if [[ "$USE_VENDOR_CACHE" == "true" && -d "$VENDOR_CACHE_DIR/vendor" ]]; then
-      echo "Using vendor cache"
-      cp -r $VENDOR_CACHE_DIR/vendor vendor
-      cp $VENDOR_CACHE_DIR/composer.lock composer.lock
-
-      # Delete magento2-base module so during composer install the package is installed again which should copy all
-      # necessary project files like app, bin, etc. There maybe a better way to do this via composer run-script but this
-      # is also works.
-      rm -rf vendor/magento/magento2-base
-    fi
-
-    if [[ ! -z "${COMPOSER_PACKAGES}" ]]; then
-      # Merge the repository object from module's composer.json into magento's composer.json
-      jq --slurpfile app $BITBUCKET_CLONE_DIR/composer.json '.repositories += [$app[0].repositories[]?]' composer.json \
-      > merged.composer.json
-
-      jq --arg path $BITBUCKET_CLONE_DIR '.repositories += [{ type: "path", "url": $path}]' merged.composer.json > composer.json
-
-      # Do not load the package we're testing from cache!
-      composer require $COMPOSER_PACKAGES "@dev" --no-update --no-cache
-    fi
-  fi
-
   composer config --no-interaction allow-plugins.dealerdirect/phpcodesniffer-composer-installer true
   composer config --no-interaction allow-plugins.laminas/laminas-dependency-plugin true
   composer config --no-interaction allow-plugins.magento/* true
 
-  cat composer.json
-  composer install
+  if [ -z "${SKIP_DEPENDENCIES}" ]; then
+    if [ ! -f "composer.lock" ]; then
+        echo "composer.lock does not exist."
+        composer create-project --repository-url="$REPOSITORY_URL" "$MAGENTO_VERSION" /magento2 --no-install
+        cd /magento2
+
+      if [[ "$USE_VENDOR_CACHE" == "true" && -d "$VENDOR_CACHE_DIR/vendor" ]]; then
+        echo "Using vendor cache"
+        cp -r $VENDOR_CACHE_DIR/vendor vendor
+        cp $VENDOR_CACHE_DIR/composer.lock composer.lock
+
+        # Delete magento2-base module so during composer install the package is installed again which should copy all
+        # necessary project files like app, bin, etc. There maybe a better way to do this via composer run-script but this
+        # is also works.
+        rm -rf vendor/magento/magento2-base
+      fi
+
+      if [[ ! -z "${COMPOSER_PACKAGES}" ]]; then
+        # Merge the repository object from module's composer.json into magento's composer.json
+        jq --slurpfile app $BITBUCKET_CLONE_DIR/composer.json '.repositories += [$app[0].repositories[]?]' composer.json \
+        > merged.composer.json
+
+        jq --arg path $BITBUCKET_CLONE_DIR '.repositories += [{ type: "path", "url": $path}]' merged.composer.json > composer.json
+
+        # Do not load the package we're testing from cache!
+        composer require $COMPOSER_PACKAGES "@dev" --no-update --no-cache
+      fi
+    fi
+    
+    cat composer.json
+    composer install
+  else 
+    echo "SKIP_DEPENDENCIES is set, so composer install will not execute."
+  fi
 
   if [[ "$USE_VENDOR_CACHE" == "true" ]]; then
     mkdir -p $VENDOR_CACHE_DIR && cp -r vendor composer.lock $VENDOR_CACHE_DIR/
@@ -76,6 +80,9 @@ composer_setup () {
 run_integration_tests () {
   composer_setup
 
+run_integration_tests () {
+  composer_install
+  
   cd dev/tests/integration
   cat etc/install-config-mysql.php.dist
